backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6